requirements.txt
selene==2.0.0rc9
pytest==8.4.1
allure-pytest==2.15.0
python-dotenv==1.1.1
selenium~=4.35.0
allure-python-commons~=2.15.0